backend/requirements.txt
fastapi
uvicorn
asyncpg
pgvector
httpx
pydantic
pydantic-settings
python-dotenv
beautifulsoup4
markdown
numpy
google-generativeai
cerebras_cloud_sdk